Moved modules command, pyglet_window, and timed_stimuli.
Command.py is the module that encapulates all that base code
to  manage the BCI commands that are passed to controllers.

Pyglet_window is the base package that manages the pyglet
runtime.

Timed_stimuli contains logic to manage emissions of values
based on time.  Such sequences can be used to create an
SSVEP M-Sequence display.